+++ Backend/src/main/java/com/backend/Java_Backend/Repository/StudentEventRepository.java
@@ -14,15 +14,15 @@ import java.util.UUID;
 @Repository
 public interface StudentEventRepository extends JpaRepository<StudentEvent, StudentEventId> {
 
-    List<StudentEvent> findByStudentId(Long studentId);
+    List<StudentEvent> findByStudentId(Integer studentId);
 
     List<StudentEvent> findByEventId(UUID eventId);
 
-    boolean existsByStudentIdAndEventId(Long studentId, UUID eventId);
+    boolean existsByStudentIdAndEventId(Integer studentId, UUID eventId);
 
     @Modifying
     @Query("DELETE FROM StudentEvent se WHERE se.student.id = :studentId AND se.event.id = :eventId")
-    void deleteByStudentIdAndEventId(@Param("studentId") Long studentId, @Param("eventId") UUID eventId);
+    void deleteByStudentIdAndEventId(@Param("studentId") Integer studentId, @Param("eventId") UUID eventId);
 
     @Query("SELECT COUNT(se) FROM StudentEvent se WHERE se.event.id = :eventId")
     int countByEventId(@Param("eventId") UUID eventId);
